Split datetime module into different files
